@@ -1,6 +1,7 @@
 {
     "name": "sgwapi",
-    "version": "1.1.1",
+    "version": "2.0.0",
+    "type": "module",
     "description": "Sendgrid Web API",
     "homepage": "https://github.com/victorenator/sgwapi",
     "author": {
@@ -8,16 +9,7 @@
         "email": "victorenator@gmail.com"
     },
     "dependencies": {
-        "debug": "^4.1.1",
-        "node-fetch": "^2.6.0"
-    },
-    "devDependencies": {
-        "@babel/core": "^7.9.0",
-        "@babel/plugin-transform-modules-commonjs": "^7.9.0",
-        "babel-plugin-add-import-extension": "^1.3.1",
-        "gulp": "^4.0.2",
-        "gulp-babel": "^8.0.0",
-        "gulp-eslint": "^6.0.0"
+        "debug": "^4.1.1"
     },
     "keywords": [
         "sendgrid",
@@ -29,6 +21,6 @@
         "url": "git://github.com/victorenator/sgwapi.git"
     },
     "engines": {
-        "node": "10"
+        "node": "^18"
     }
 }
